Repo structure cleanup
Addressing issue #100
Reference: https://www.youtube.com/watch?v=VYycQTm2HrM
