+++ Pipfile
@@ -29,7 +29,7 @@ mkdocs-material = "*"
 python_version = "3.9"
 
 [scripts]
-run = "python runner.py"
+run = "python src/runner.py"
 quick = "python -m streamlit run interface.py"
 init = "git submodule update --init --remote --recursive"
 update = "git submodule update --remote --recursive"
