--- docs/advanced-usage.md
@@ -1,3 +1,0 @@
-# Advanced Usage
-
-Coming soon.
