--- docs/contributing.md
@@ -1,5 +1,0 @@
-# Contribution
-
-`Sammlung Toole` is free, open source software. All contributions are welcome.
-
-More informations to follow soon. For now, please just get in touch through a GitHub issue.
