--- docs/development.md
@@ -1,3 +1,0 @@
-# Development
-
-Coming soon. For now, please consult the [readme](https://github.com/arbeitsgruppe-digitale-altnordistik/Sammlung-Toole#readme).
