--- docs/index.md
@@ -1,11 +1,0 @@
-# Sammlung Toole
-
-Welcome to Ultima Toole!
-
-`Sammlung Toole` is a project by the "Arbeitsgruppe Digitale Arbeitsmethoden in der Altnordistik" at the University of Basel.
-
-The tool aims to shed new light on the data found on [Handrit.is](https://handrit.is/), beyond what their user interface can show.
-
-It is written in Python using the UI framework Streamlit.
-
-For any information on how to use it etc., please consult [this documentation](https://arbeitsgruppe-digitale-altnordistik.github.io/Sammlung-Toole/).
