--- docs/simple/browse.md
@@ -1,5 +1,0 @@
-# Browse Data
-
-To get an impression of the data loaded by the tool, you can visit the `Browse Data` pare.
-
-Here you will find information on the manuscripts, texts and people, as well as seeing a list of all of them.
