--- docs/simple/groups.md
@@ -1,3 +1,0 @@
-# Groups
-
-Coming soon.
