--- docs/simple/search.md
@@ -1,47 +1,0 @@
-# Search Functions
-
-## What to search for
-
-There are multiple search options that can be selected through the radio buttons in the sidebar:
-
-- Handrit URLs:  
-  Use one/multiple search- or browse result URL from Handrit.is.  
-  The tool will find the manuscripts as returned by Handrit.is, and show the according metadata.  
-  This is handy, if you have already found an interestiong set if manuscripts while searching on Handrit.is, and want to get the same manuscripts displayed in the tool.
-
-
-- Manuscript by Person:  
-  Select one/multiple persons form the Handrit.is authority file.  
-  The tool will find all manuscripts related to one/all of the selected people.
-
-
-- Person by Manuscript:  
-  Select one/multiple manuscripts form the Handrit.is collection.  
-  The tool will find all people related to one/all of the selected manuscripts.
-
-
-- Manuscript by Text:  
-  Select one/multiple texts mentioned in the Handrit.is collections.  
-  The tool will find all manuscripts related to one/all of the selected texts.
-
-
-- Text by Manuscript:  
-  Select one/multiple manuscripts form the Handrit.is collection.  
-  The tool will find all texts occuring in one/all of the selected manuscripts.
-
-
-## Search Mode: Union/Intersection, AND/OR
-
-For each search type, it's possible to search for one or multiple entities (e.g. multiple people, search links or shelfmarks).  
-For each search, it's also necessary to pick a search mode. The two options are:
-
-- `AND` (Intersection)
-- `OR` (Union)
-
-When searching for only one entity, the search mode does not make any difference.  
-When searching for two or more entities, the search mode determines, if
-
-- a manuscript must match *one (or more)* of the entities searched for (union)
-- a manuscript must match *all* of the entities searched for (intersection)
-
-in order to match the search criteria.
