--- docs/simple/usage-overview.md
@@ -1,22 +1,0 @@
-# Usage Overview
-
-
-## Where to find Sammlung Toole?
-
-Currently, no public server is available. For access to the *super secret* internal server, please get in touch with us.
-
-Alternatively, you can run the tool locally on your computer, by following [these instructions](../advanced-usage.md).
-
-
-## How to use Sammlung Toole
-
-Once you have access to the tool (be that on a server or locally), simply open the URL in your browser.
-
-On normal screen sizes, you will see the navigation bar on the left side on the page.
-
-> If you don't see the navigation bar, there is a chance it's just collapsed/hidden. Look out for a little arrow on the very left edge of the screen.
-
-
-> On small screens, especially on mobile devices, the layout may look slightly different.
-
-In the dropdown located in the navigation bar, you can simply select a page (e.g. [Groups](groups.md) or [Search](search.md)), and the application will take you there.
