--- docs/simple/visualization.md
@@ -1,3 +1,0 @@
-# Visualization
-
-Coming soon.
