--- gui/__init__.py
@@ -1,2 +1,0 @@
-from .pages import groups
-from .pages import search
